Revert float stock

+++ src/Pyz/Yves/QuickOrderPage/Theme/default/components/molecules/quick-order-row/quick-order-row.ts
@@ -92,7 +92,7 @@ export default class QuickOrderRow extends Component {
 
     async reloadField(sku: string = '') {
         clearTimeout(this.timer);
-        const quantityInputValue = parseFloat(this.quantityValue);
+        const quantityInputValue = parseInt(this.quantityValue);
 
         this.ajaxProvider.queryParams.set('sku', sku);
         this.ajaxProvider.queryParams.set('index', this.ajaxProvider.getAttribute('class').split('-').pop().trim());
